refactor/product-detail
